+++ packages/language-server/test/unit/completion-items-classes.test.ts
@@ -6,10 +6,13 @@ import {
 } from "vscode-languageserver";
 import { UI5SemanticModel } from "@ui5-language-assistant/semantic-model-types";
 import {
+  DEFAULT_UI5_VERSION,
+  DEFAULT_UI5_FRAMEWORK,
+} from "@ui5-language-assistant/constant";
+import {
   generateModel,
   expectExists,
   getFallbackPatchVersions,
-  DEFAULT_UI5_VERSION,
 } from "@ui5-language-assistant/test-utils";
 import { generate } from "@ui5-language-assistant/semantic-model";
 import {
@@ -27,7 +30,7 @@ describe("the UI5 language assistant Code Completion Services - classes", () => 
   let appContext: AppContext;
   beforeAll(async function () {
     ui5SemanticModel = await generateModel({
-      framework: "SAPUI5",
+      framework: DEFAULT_UI5_FRAMEWORK,
       version: (
         await getFallbackPatchVersions()
       ).SAPUI5 as typeof DEFAULT_UI5_VERSION,