+++ packages/language-server/test/unit/hover.test.ts
@@ -8,10 +8,13 @@ import {
 } from "vscode-languageserver";
 import { UI5SemanticModel } from "@ui5-language-assistant/semantic-model-types";
 import {
+  DEFAULT_UI5_FRAMEWORK,
+  DEFAULT_UI5_VERSION,
+} from "@ui5-language-assistant/constant";
+import {
   generateModel,
   expectExists,
   getFallbackPatchVersions,
-  DEFAULT_UI5_VERSION,
 } from "@ui5-language-assistant/test-utils";
 import { generate } from "@ui5-language-assistant/semantic-model";
 import { getHoverResponse } from "../../src/hover";
@@ -24,7 +27,7 @@ describe("the UI5 language assistant Hover Tooltip Service", () => {
   let appContext: AppContext;
   beforeAll(async () => {
     ui5SemanticModel = await generateModel({
-      framework: "SAPUI5",
+      framework: DEFAULT_UI5_FRAMEWORK,
       version: (
         await getFallbackPatchVersions()
       ).SAPUI5 as typeof DEFAULT_UI5_VERSION,