+++ packages/xml-views-completion/test/unit/providers/elementName/aggregation.test.ts
@@ -2,7 +2,10 @@ import { map, cloneDeep, forEach } from "lodash";
 import { XMLElement } from "@xml-tools/ast";
 import { UI5SemanticModel } from "@ui5-language-assistant/semantic-model-types";
 import {
+  DEFAULT_UI5_FRAMEWORK,
   DEFAULT_UI5_VERSION,
+} from "@ui5-language-assistant/constant";
+import {
   buildUI5Aggregation,
   generateModel,
   getFallbackPatchVersions,
@@ -18,7 +21,7 @@ describe("The ui5-language-assistant xml-views-completion", () => {
   let appContext: AppContext;
   beforeAll(async () => {
     REAL_UI5_MODEL = await generateModel({
-      framework: "SAPUI5",
+      framework: DEFAULT_UI5_FRAMEWORK,
       version: (
         await getFallbackPatchVersions()
       ).SAPUI5 as typeof DEFAULT_UI5_VERSION,