+++ packages/xml-views-tooltip/test/unit/utils.ts
@@ -1,5 +1,6 @@
 import { UI5SemanticModel } from "@ui5-language-assistant/semantic-model-types";
 import { Context } from "@ui5-language-assistant/context";
+import { DEFAULT_UI5_FRAMEWORK } from "@ui5-language-assistant/constant";
 
 export const getDefaultContext = (ui5Model: UI5SemanticModel): Context => {
   return {
@@ -15,7 +16,7 @@ export const getDefaultContext = (ui5Model: UI5SemanticModel): Context => {
     },
     services: {},
     yamlDetails: {
-      framework: "SAPUI5",
+      framework: DEFAULT_UI5_FRAMEWORK,
       version: undefined,
     },
   };