+++ packages/xml-views-validation/test/unit/validators/element/type-of-aggregation.test.ts
@@ -4,7 +4,10 @@ import {
   UI5Aggregation,
 } from "@ui5-language-assistant/semantic-model-types";
 import {
+  DEFAULT_UI5_FRAMEWORK,
   DEFAULT_UI5_VERSION,
+} from "@ui5-language-assistant/constant";
+import {
   generateModel,
   getFallbackPatchVersions,
 } from "@ui5-language-assistant/test-utils";
@@ -28,7 +31,7 @@ describe("the type aggregation validation", () => {
   let appContext: AppContext;
   beforeAll(async () => {
     ui5SemanticModel = await generateModel({
-      framework: "SAPUI5",
+      framework: DEFAULT_UI5_FRAMEWORK,
       version: (
         await getFallbackPatchVersions()
       ).SAPUI5 as typeof DEFAULT_UI5_VERSION,