+++ packages/vscode-ui5-language-assistant/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Change Log
 
+## 4.0.69
+
+### Patch Changes
+
+- 65eaef5: fix: change SAP UI5 to SAPUI5
+
+## 4.0.68
+
+### Patch Changes
+
+- a82d24c: fix: reference
+- Updated dependencies [a82d24c]
+  - @ui5-language-assistant/language-server@4.0.52
+
+## 4.0.67
+
+### Patch Changes
+
+- b2a8d01: fix: introduce constant package to handle UI5 default version in central package
+- Updated dependencies [b2a8d01]
+  - @ui5-language-assistant/language-server@4.0.51
+
 ## 4.0.66
 
 ### Patch Changes